move out compute command to separate module
